Add CheckAndSetExceptionMapper for error handling
